Make it work on Harmattan.
The application now runs on Harmattan. but there are still a number
of issues. I had to disable the overview, because it crashed the
application. There are also some non-working stubs in the microkde lib,
which is a thin layer between Qt and Polka making it run without
the KDE libraries. The UI needs work. But Polka runs and can be
operated on a real phone now.