Remove Meego port
Meego is obsolete and after porting the code to KDE Frameworks 5
it doesn't work anymore. So instead of trying to fix it let's just
remove it.

The concept is still valid and nice and also applicable to other
platforms where you only want to rely on Qt itself as a dependency.
With KDE Framekworks 5 this is much less relevant than before, though,
as the needed frameworks are working or can be made to work more
easily.